+++ changelog.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- [Add a file content listener to read the content of a file once it is added to the remote file system](https://github.com/ballerina-platform/ballerina-library/issues/1490)
+
 ### Fixed
 
 - [Fix the issue where the FTP URL is improperly formatted for the jail-break scnenarios](https://github.com/ballerina-platform/ballerina-library/issues/8267)